refactor: GameController, TeamController 분리
